+++ CHANGELOG.md
@@ -7,6 +7,8 @@
         - fix panic for invalid input.
         - add gaps of `k-1` bp before concatatenating seqs.
     - add warning for invalid input.
+- `profile`:
+    - *allow changing parts of the parameters for preset profiling modes*. [#5](https://github.com/shenwei356/kmcp/issues/5)
 
 ### v0.8.1 - 2022-03-07